Moved class Publication to mpcosarchive
